fix: reorganized profiles and user_directory
